added option to train both, added learning rate as command line argument

+++ train.py
@@ -19,7 +19,8 @@ import optim
 import graph
 
 
-def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
+def experiment(n_ins=[7,16],n_hiddens=[64,256],overwrite=False,
+               learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                l1_reg=0, l2_reg=1e-3, n_epochs=500, init_patience=20, 
                batch_size=1000, repeated_exp=5,
                to_run=['gradclip','ortho3','both']):
@@ -44,7 +45,7 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                 (xreshape(test_set_x),test_set_y)]
     
     # Define all the experiment functions
-    def test_gradclip(clip_limit,n_in,n_hidden,seed=1):
+    def test_gradclip(clip_limit,n_in,n_hidden,learning_rate,seed=1):
         # Build the model
         print('... building the gradient-clipped RNN')
         x = T.tensor3('x')
@@ -58,7 +59,7 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                          learning_rate,lr_decay,momentum,
                          init_patience,n_epochs)
         
-    def test_orthopen3(eps_ortho,n_in,n_hidden,seed=1):
+    def test_orthopen3(eps_ortho,n_in,n_hidden,learning_rate,seed=1):
         # Build the model
         print('... building the 3rd orthogonality-constrained RNN')
         x = T.tensor3('x')
@@ -67,6 +68,21 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
         model = recurrent.rnn_ortho3(x,n_in,n_hidden,10,bptt_limit,rng)
         cost = model.crossentropy(y)+l1_reg*model.L1+l2_reg*model.L2+ \
                eps_ortho*model.ortho
+        return optim.sgd(dataprep(n_in),model,cost,x,y,n_train_batches,
+                         n_valid_batches,n_test_batches,batch_size,
+                         learning_rate,lr_decay,momentum,
+                         init_patience,n_epochs)
+
+    def test_both(eps_ortho,clip_limit,n_in,n_hidden,learning_rate,seed=1):
+        # Build the model
+        print('... building orthogonality-constrained gradient-clipped RNN')
+        x = T.tensor3('x')
+        y = T.ivector('y')
+        rng = numpy.random.RandomState(seed)
+        model = recurrent.rnn_ortho3(x,n_in,n_hidden,10,bptt_limit,rng)
+        unclipped_cost = model.crossentropy(y)+l1_reg*model.L1 \
+                         +l2_reg*model.L2+eps_ortho*model.ortho
+        cost = theano.gradient.grad_clip(unclipped_cost,-clip_limit,clip_limit)
         return optim.sgd(dataprep(n_in),model,cost,x,y,n_train_batches,
                          n_valid_batches,n_test_batches,batch_size,
                          learning_rate,lr_decay,momentum,
@@ -91,7 +107,7 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
     
     # log results in a csv file
     def log_results(filename,line,hyperparam,n_in,n_hidden,
-                    trn,vld,tst,mntrs):
+                    trn,vld,tst,mntrs,overwrite=False):
         import csv
         import os
         # unpack the results
@@ -103,7 +119,7 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
         normWh = [n[1] for n in mntrs]
         normWy = [n[2] for n in mntrs]
         orthog = [n[3] for n in mntrs]
-        if line==0:
+        if line==0 and overwrite:
             # check if old log exists and delete
             if os.path.isfile(filename):
                 os.remove(filename)
@@ -120,10 +136,6 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                          vld_loss,vld_error,tst,
                          normWx,normWh,normWy,orthog))
     
-    # experiment parameters
-    n_ins = [7,16]
-    n_hiddens = [64,256]
-    
     # hyperparameter search for gradient clipping
     if any('gradclip' in s for s in to_run):
         clip_limits = [10**x for x in range(-3,2)]
@@ -131,10 +143,11 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                 enumerate(itertools.product(clip_limits,
                                             n_ins,
                                             n_hiddens)):
-            fn = lambda seed: test_gradclip(clip_limit,n_in,n_hidden,seed)
+            fn = lambda seed: test_gradclip(clip_limit,n_in,n_hidden,
+                                            learning_rate,seed)
             trn,vld,tst,mntrs = repeat_test(fn,repeated_exp)
             log_results('gradclip.csv',idx,clip_limit,n_in,n_hidden,
-                        trn,vld,tst,mntrs)
+                        trn,vld,tst,mntrs,overwrite)
         graph.make_all('gradclip.csv')
     
     # hyperparameter search for orthogonality penalty
@@ -144,16 +157,45 @@ def experiment(learning_rate=1e-1, lr_decay=0.99, bptt_limit=784, momentum=0.9,
                 enumerate(itertools.product(eps_orthos,
                                             n_ins,
                                             n_hiddens)):
-            fn = lambda seed: test_orthopen3(eps_ortho,n_in,n_hidden,seed)
+            fn = lambda seed: test_orthopen3(eps_ortho,n_in,n_hidden,
+                                             learning_rate,seed)
             trn,vld,tst,mntrs = repeat_test(fn,repeated_exp)
             log_results('orthopen3.csv',idx,eps_ortho,n_in,n_hidden,
-                        trn,vld,tst,mntrs)
+                        trn,vld,tst,mntrs,overwrite)
         graph.make_all('orthopen3.csv')
+    
+    # hyperparameter search for combination model
+    if any('both' in s for s in to_run):
+        eps_orthos = [10**x for x in range(-5,0)]
+        clip_limits = [10**x for x in range(-3,2)]
+        for idx,(eps_ortho,clip_limit,n_in,n_hidden) in \
+                enumerate(itertools.product(eps_orthos,
+                                            clip_limits,
+                                            n_ins,
+                                            n_hiddens)):
+            fn = lambda seed: test_both(eps_ortho,clip_limit,n_in,n_hidden,
+                                        learning_rate,seed)
+            trn,vld,tst,mntrs = repeat_test(fn,repeated_exp)
+            log_results('both.csv',idx,[eps_ortho,clip_limit],n_in,n_hidden,
+                        trn,vld,tst,mntrs,overwrite)
+        graph.make_all('both.csv')
 
 if __name__ == "__main__":
     import argparse
     parser = argparse.ArgumentParser(description='Run RNN experiments')
     parser.add_argument('--exp',nargs='*',type=str,
-                        default=['gradclip','ortho1','ortho2','ortho3'])
+                        default=['gradclip','ortho3','both'])
+    parser.add_argument('--n_in',nargs='*',type=int,
+                        default=[7,16])
+    parser.add_argument('--n_hidden',nargs='*',type=int,
+                        default=[64,256])
+    parser.add_argument('--learnrate',nargs='*',type=float,
+                        default=0.1)
+    parser.add_argument('--overwrite',nargs='*',type=bool,
+                        default=False)
     print(parser.parse_args().exp)
-    experiment(to_run=parser.parse_args().exp)
+    experiment(to_run=parser.parse_args().exp,
+               n_ins=parser.parse_args().n_in,
+               n_hiddens=parser.parse_args().n_hidden,
+               learning_rate=parser.parse_args().learnrate,
+               overwrite=parser.parse_args().overwrite)
